meta-aws-iot-base/aws-cdk/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
attrs==25.3.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.227
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.1.0
    # via aws-cdk-lib
aws-cdk-aws-iot-actions-alpha==2.183.0a0
    # via aws_cdk_iot (setup.py)
aws-cdk-aws-iot-alpha==2.183.0a0
    # via
    #   aws-cdk-aws-iot-actions-alpha
    #   aws_cdk_iot (setup.py)
aws-cdk-aws-iotevents-alpha==2.183.0a0
    # via aws-cdk-aws-iot-actions-alpha
aws-cdk-aws-kinesisfirehose-alpha==2.183.0a0
    # via
    #   aws-cdk-aws-iot-actions-alpha
    #   aws-cdk-aws-kinesisfirehose-destinations-alpha
aws-cdk-aws-kinesisfirehose-destinations-alpha==2.183.0a0
    # via aws-cdk-aws-iot-actions-alpha
aws-cdk-cloud-assembly-schema==40.7.0
    # via aws-cdk-lib
aws-cdk-lib==2.183.0
    # via
    #   aws-cdk-aws-iot-actions-alpha
    #   aws-cdk-aws-iot-alpha
    #   aws-cdk-aws-iotevents-alpha
    #   aws-cdk-aws-kinesisfirehose-alpha
    #   aws-cdk-aws-kinesisfirehose-destinations-alpha
    #   aws_cdk_iot (setup.py)
cattrs==24.1.2
    # via jsii
constructs==10.4.2
    # via
    #   aws-cdk-aws-iot-actions-alpha
    #   aws-cdk-aws-iot-alpha
    #   aws-cdk-aws-iotevents-alpha
    #   aws-cdk-aws-kinesisfirehose-alpha
    #   aws-cdk-aws-kinesisfirehose-destinations-alpha
    #   aws-cdk-lib
    #   aws_cdk_iot (setup.py)
exceptiongroup==1.2.2
    # via cattrs
importlib-resources==6.5.2
    # via jsii
jsii==1.109.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-iot-actions-alpha
    #   aws-cdk-aws-iot-alpha
    #   aws-cdk-aws-iotevents-alpha
    #   aws-cdk-aws-kinesisfirehose-alpha
    #   aws-cdk-aws-kinesisfirehose-destinations-alpha
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-iot-actions-alpha
    #   aws-cdk-aws-iot-alpha
    #   aws-cdk-aws-iotevents-alpha
    #   aws-cdk-aws-kinesisfirehose-alpha
    #   aws-cdk-aws-kinesisfirehose-destinations-alpha
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
    #   jsii
python-dateutil==2.9.0.post0
    # via jsii
six==1.17.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-iot-actions-alpha
    #   aws-cdk-aws-iot-alpha
    #   aws-cdk-aws-iotevents-alpha
    #   aws-cdk-aws-kinesisfirehose-alpha
    #   aws-cdk-aws-kinesisfirehose-destinations-alpha
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.12.2
    # via
    #   cattrs
    #   jsii
